Changed all stores, but .js files won't compile.
